# www/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Fiat Lux api documentation",
        "description": "Fiat Lux School OpenApi documentation",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Demo API Server"
        }
    ],
    "paths": {
        "/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Login",
                "description": "Get a JWT via given credentials.",
                "operationId": "67d6d0fdd54fba132e1b6f55ca254183",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "access_token": {
                                            "type": "string"
                                        },
                                        "token_type": {
                                            "type": "string"
                                        },
                                        "expires_in": {
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden access"
                    },
                    "422": {
                        "description": "Wrong credentials"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                }
            }
        },
        "/logout": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Logout",
                "description": "Log the user out (Invalidate the token).",
                "operationId": "28d48c13fca984400ac8de99f54ee13a",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthorized access"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/tests/{testId}/options": {
            "post": {
                "tags": [
                    "Options"
                ],
                "summary": "Saves question option",
                "description": "Store a newly created resource in storage.",
                "operationId": "36b1d193e7cfde0631695ef7b487740e",
                "parameters": [
                    {
                        "name": "testId",
                        "in": "path",
                        "description": "Question id",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreOptionRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Option"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized operation"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/tests/{testId}/options/{optionId}": {
            "post": {
                "tags": [
                    "Options"
                ],
                "summary": "Updates question option",
                "description": "Update the specified resource in storage.",
                "operationId": "337d9fccf31ec5dd2477b46227ba2fac",
                "parameters": [
                    {
                        "name": "testId",
                        "in": "path",
                        "description": "Question id",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    },
                    {
                        "name": "optionId",
                        "in": "path",
                        "description": "Option id",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreOptionRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Option"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized operation"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "StoreOptionRequest": {
                "title": "Store Option request",
                "description": "Store Option request body data",
                "required": [
                    "text",
                    "is_correct"
                ],
                "properties": {
                    "text": {
                        "title": "Option text",
                        "description": "Option text",
                        "type": "string"
                    },
                    "is_correct": {
                        "title": "Is correct option",
                        "description": "Flag for correct test option",
                        "type": "boolean"
                    }
                },
                "type": "object"
            },
...
            "Option": {
                "title": "Option Response model",
                "description": "Option Response model",
                "properties": {
                    "id": {
                        "title": "ID",
                        "description": "Current test id",
                        "type": "integer",
                        "format": "int64"
                    },
                    "text": {
                        "title": "Option text",
                        "description": "Option text",
                        "type": "string",
                        "format": "string"
                    },
                    "is_correct": {
                        "title": "Is correct option flag",
                        "description": "MArk if current option is correct for the question",
                        "type": "boolean",
                        "format": "boolean"
                    },
                    "test_id": {
                        "title": "Test id",
                        "description": "Test id for the current option",
                        "type": "integer",
                        "format": "int64"
                    },
                    "test": {
                        "$ref": "#/components/schemas/Test"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Option"
                }
            },
            "Test": {
                "title": "Test Response model",
                "description": "Test Response model",
                "properties": {
                    "id": {
                        "title": "ID",
                        "description": "Current test id",
                        "type": "integer",
                        "format": "int64"
                    },
                    "question": {
                        "title": "Test question",
                        "description": "Test question",
                        "type": "string",
                        "format": "string"
                    },
                    "test_group_id": {
                        "title": "Test group id",
                        "description": "Test group id",
                        "type": "integer",
                        "format": "int64"
                    },
                    "test_group": {
                        "$ref": "#/components/schemas/TestGroup"
                    },
                    "options": {
                        "title": "List of options",
                        "description": "List of options for the current test",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Option"
                        }
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Test"
                }
            },
            "TestGroup": {
                "title": "Test Group Response model",
                "description": "Test Group Response model",
                "properties": {
                    "id": {
                        "title": "ID",
                        "description": "Current test group id",
                        "type": "integer",
                        "format": "int64"
                    },
                    "group_name": {
                        "title": "Test group name",
                        "description": "Test group name",
                        "type": "string",
                        "format": "string"
                    },
                    "tests": {
                        "title": "Questions list",
                        "description": "List of test questions",
                        "type": "array",
                        "format": "array",
                        "items": {
                            "$ref": "#/components/schemas/Test"
                        }
                    }
                },
                "type": "object",
                "xml": {
                    "name": "TestGroup"
                }
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "Bearer",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}